Unifies admin image selection and upload flow
Replaces ad hoc file-prompt uploads with a shared image picker so admins can reuse existing assets or upload new ones in one place.

Improves editing UX by adding image previews and clear actions, and exposes image library access from the admin navigation for faster asset management.

Centralizes upload handling through a common admin image pipeline and updates database typing to support stored image assets.

Fixes date recalculation SQL by enforcing integer day offsets, preventing type-related date arithmetic issues.

+++ sql/fix-vinyl-sundays-series-7-15.sql
@@ -20,7 +20,7 @@ ORDER BY id;
 UPDATE events
 SET
   date = (
-    DATE '2026-03-29' + ((id - 9) * 7)
+    DATE '2026-03-29' + (((id - 9) * 7)::int)
   )::date,
   has_queue = true,
   -- Keep existing queue_types when present, otherwise provide a safe default.
